refactor: OwlChatBubble・TodoCardを共通コンポーネントに抽出
重複していたフクロウ吹き出しとチェックリストのコードを
共通コンポーネントに抽出し、両画面から利用するよう変更
